Counting stack
In this simple exercise I create a class which mimicks a stack, and give it push and pop instance methods. I then extend the implementation by creating another (sub)class which inherits from the 'Stack' class. The subclass overrides the super class pop method and counts the number of pops. I create another instance method to return the count.